+++ bot/src/test/java/edu/java/bot/command/HelpCommandTest.java
@@ -5,6 +5,7 @@ import com.pengrad.telegrambot.model.Message;
 import com.pengrad.telegrambot.model.Update;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
+import edu.java.bot.HelpCommand;
 import edu.java.bot.service.UserService;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeAll;
